Implement FtpServer.
