+++ MsSqlFtpServer/MembershipProviderType.cs
@@ -1,0 +1,19 @@
+﻿namespace MsSqlFtpServer
+{
+    /// <summary>
+    /// The selected membership provider.
+    /// </summary>
+    public enum MembershipProviderType
+    {
+        /// <summary>
+        /// Use the custom (example) membership provider.
+        /// </summary>
+        Custom,
+
+        /// <summary>
+        /// Use the membership provider for anonymous users.
+        /// </summary>
+        Anonymous,
+    }
+
+}
