+++ MsSqlFtpServer/MsSqlFtpServer.csproj
@@ -33,6 +33,18 @@
     <WarningLevel>4</WarningLevel>
   </PropertyGroup>
   <ItemGroup>
+    <Reference Include="FubarDev.FtpServer, Version=3.0.0.0, Culture=neutral, processorArchitecture=MSIL">
+      <HintPath>..\packages\FubarDev.FtpServer.3.0.0-rc.2\lib\net47\FubarDev.FtpServer.dll</HintPath>
+    </Reference>
+    <Reference Include="FubarDev.FtpServer.Abstractions, Version=3.0.0.0, Culture=neutral, processorArchitecture=MSIL">
+      <HintPath>..\packages\FubarDev.FtpServer.Abstractions.3.0.0-rc.2\lib\net461\FubarDev.FtpServer.Abstractions.dll</HintPath>
+    </Reference>
+    <Reference Include="FubarDev.FtpServer.Commands, Version=3.0.0.0, Culture=neutral, processorArchitecture=MSIL">
+      <HintPath>..\packages\FubarDev.FtpServer.Commands.3.0.0-rc.2\lib\net461\FubarDev.FtpServer.Commands.dll</HintPath>
+    </Reference>
+    <Reference Include="Microsoft.AspNetCore.Http.Features, Version=2.0.0.0, Culture=neutral, PublicKeyToken=adb9793829ddae60, processorArchitecture=MSIL">
+      <HintPath>..\packages\Microsoft.AspNetCore.Http.Features.2.0.0\lib\netstandard2.0\Microsoft.AspNetCore.Http.Features.dll</HintPath>
+    </Reference>
     <Reference Include="Microsoft.CSharp" />
     <Reference Include="Microsoft.DotNet.PlatformAbstractions, Version=2.1.0.0, Culture=neutral, PublicKeyToken=adb9793829ddae60, processorArchitecture=MSIL">
       <HintPath>..\..\FtpServer\packages\Microsoft.DotNet.PlatformAbstractions.2.1.0\lib\net45\Microsoft.DotNet.PlatformAbstractions.dll</HintPath>
@@ -69,6 +81,9 @@
     </Reference>
     <Reference Include="Newtonsoft.Json, Version=9.0.0.0, Culture=neutral, PublicKeyToken=30ad4fe6b2a6aeed, processorArchitecture=MSIL">
       <HintPath>..\..\FtpServer\packages\Newtonsoft.Json.9.0.1\lib\net45\Newtonsoft.Json.dll</HintPath>
+    </Reference>
+    <Reference Include="NGettext, Version=0.6.1.0, Culture=neutral, PublicKeyToken=08d3d1c89dfd2985, processorArchitecture=MSIL">
+      <HintPath>..\packages\NGettext.0.6.3\lib\net46\NGettext.dll</HintPath>
     </Reference>
     <Reference Include="NLog, Version=4.0.0.0, Culture=neutral, PublicKeyToken=5120e14c03d0593c, processorArchitecture=MSIL">
       <HintPath>..\..\FtpServer\packages\NLog.4.6.3\lib\net45\NLog.dll</HintPath>
@@ -107,6 +122,12 @@
     </Reference>
     <Reference Include="System.Runtime.Serialization" />
     <Reference Include="System.ServiceModel" />
+    <Reference Include="System.Threading.Channels, Version=4.0.0.0, Culture=neutral, PublicKeyToken=cc7b13ffcd2ddd51, processorArchitecture=MSIL">
+      <HintPath>..\packages\System.Threading.Channels.4.5.0\lib\netstandard2.0\System.Threading.Channels.dll</HintPath>
+    </Reference>
+    <Reference Include="System.Threading.Tasks.Extensions, Version=4.2.0.0, Culture=neutral, PublicKeyToken=cc7b13ffcd2ddd51, processorArchitecture=MSIL">
+      <HintPath>..\packages\System.Threading.Tasks.Extensions.4.5.0\lib\netstandard2.0\System.Threading.Tasks.Extensions.dll</HintPath>
+    </Reference>
     <Reference Include="System.Transactions" />
     <Reference Include="System.Xml.Linq" />
     <Reference Include="System.Data.DataSetExtensions" />
@@ -115,6 +136,8 @@
     <Reference Include="System.Xml" />
   </ItemGroup>
   <ItemGroup>
+    <Compile Include="FtpServerConfigOptions.cs" />
+    <Compile Include="MembershipProviderType.cs" />
     <Compile Include="Program.cs" />
     <Compile Include="Properties\AssemblyInfo.cs" />
   </ItemGroup>
@@ -133,18 +156,6 @@
     <None Include="packages.config" />
   </ItemGroup>
   <ItemGroup>
-    <ProjectReference Include="..\..\FtpServer\src\FubarDev.FtpServer.Abstractions\FubarDev.FtpServer.Abstractions.csproj">
-      <Project>{37ce4218-fac3-4201-82a0-54945698c6ad}</Project>
-      <Name>FubarDev.FtpServer.Abstractions</Name>
-    </ProjectReference>
-    <ProjectReference Include="..\..\FtpServer\src\FubarDev.FtpServer.Commands\FubarDev.FtpServer.Commands.csproj">
-      <Project>{7e3f81e2-764e-4075-992f-e53bf50f9623}</Project>
-      <Name>FubarDev.FtpServer.Commands</Name>
-    </ProjectReference>
-    <ProjectReference Include="..\..\FtpServer\src\FubarDev.FtpServer\FubarDev.FtpServer.csproj">
-      <Project>{01ebbf94-7469-43a0-a618-07fb734b163f}</Project>
-      <Name>FubarDev.FtpServer</Name>
-    </ProjectReference>
     <ProjectReference Include="..\FtpServer.MsSqlFileSystem\FtpServer.MsSqlFileSystem.csproj">
       <Project>{2b8c8e93-feec-433a-aaa9-8000d54bca05}</Project>
       <Name>FtpServer.MsSqlFileSystem</Name>
