+++ MsSqlFtpServer/packages.config
@@ -1,5 +1,9 @@
 ﻿<?xml version="1.0" encoding="utf-8"?>
 <packages>
+  <package id="FubarDev.FtpServer" version="3.0.0-rc.2" targetFramework="net472" />
+  <package id="FubarDev.FtpServer.Abstractions" version="3.0.0-rc.2" targetFramework="net472" />
+  <package id="FubarDev.FtpServer.Commands" version="3.0.0-rc.2" targetFramework="net472" />
+  <package id="Microsoft.AspNetCore.Http.Features" version="2.0.0" targetFramework="net472" />
   <package id="Microsoft.DotNet.PlatformAbstractions" version="2.1.0" targetFramework="net472" />
   <package id="Microsoft.Extensions.Configuration" version="3.0.0-preview5.19227.9" targetFramework="net472" />
   <package id="Microsoft.Extensions.Configuration.Abstractions" version="3.0.0-preview5.19227.9" targetFramework="net472" />
@@ -12,6 +16,7 @@
   <package id="Microsoft.Extensions.Options" version="3.0.0-preview5.19227.9" targetFramework="net472" />
   <package id="Microsoft.Extensions.Primitives" version="3.0.0-preview5.19227.9" targetFramework="net472" />
   <package id="Newtonsoft.Json" version="9.0.1" targetFramework="net472" />
+  <package id="NGettext" version="0.6.3" targetFramework="net472" />
   <package id="NLog" version="4.6.3" targetFramework="net472" />
   <package id="NLog.Extensions.Logging" version="1.5.0" targetFramework="net472" />
   <package id="Scrutor" version="2.0.0" targetFramework="net472" />
@@ -21,4 +26,6 @@
   <package id="System.Numerics.Vectors" version="4.4.0" targetFramework="net472" />
   <package id="System.Runtime.CompilerServices.Unsafe" version="4.6.0-preview5.19224.8" targetFramework="net472" />
   <package id="System.Runtime.InteropServices.RuntimeInformation" version="4.0.0" targetFramework="net472" />
+  <package id="System.Threading.Channels" version="4.5.0" targetFramework="net472" />
+  <package id="System.Threading.Tasks.Extensions" version="4.5.0" targetFramework="net472" />
 </packages>
